wip: heaps of refactoring and fixes for the PR review
