+++ src/backends/fvm.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "fvm_multicore.hpp"
+#include "./multicore/fvm.hpp"
 
 #ifdef NMC_HAVE_CUDA
-    #include "fvm_gpu.hpp"
+    #include "./gpu/fvm.hpp"
 #endif
