+++ src/backends/gpu/intrinsics.hpp
