+++ src/backends/matrix_storage.md
@@ -76,5 +76,3 @@ p_flt = [0 0 1 2 1 4 | 6 6 7 8 7 10 11 10]
 
 p_int = [0 1 * *| 0 1 * * | 4 5 * * | 8 9 * * | 4 5 * * | 16 17 * * | 20 * * * | 16 * * * ]
 
-
-
