+++ src/backends/multicore/stimulus.hpp
