+++ tests/unit/test_matrix.cpp
@@ -5,7 +5,7 @@
 
 #include <math.hpp>
 #include <matrix.hpp>
-#include <backends/fvm_multicore.hpp>
+#include <backends/multicore/fvm.hpp>
 #include <util/span.hpp>
 
 using namespace nest::mc;
